Don't call the modules API if the project specifies a target.
This is a bit confusing because the Django settings is called
WORKER_MODULE_NAME, but the function will return the version name
by default.

I think the version name is more useful - it will cause the tasks to
be run on the same module / version as that of the original request.